+++ app_estudo/integrations/__init__.py
@@ -2,10 +2,13 @@
 
 from .ankiconnect_healthcheck import AnkiHealthcheckResult, check_ankiconnect
 from .anki_mapping import LogicalAnkiNote, map_to_anki_logical_note
+from .ankiconnect_client import AnkiConnectClient, AnkiSyncResult
 
 __all__ = [
 	"LogicalAnkiNote",
 	"map_to_anki_logical_note",
 	"AnkiHealthcheckResult",
 	"check_ankiconnect",
+	"AnkiConnectClient",
+	"AnkiSyncResult",
 ]